+++ src/client/java/io/github/drakonkinst/worldsinger/datagen/ModTextureKeys.java
@@ -27,7 +27,7 @@ import net.minecraft.client.data.TextureKey;
 
 public final class ModTextureKeys {
 
-    public static final TextureKey ALUMINUM = TextureKey.of("aluminum");
+    public static final TextureKey SIDE_INNER = TextureKey.of("side_inner");
     public static final TextureKey CONTENTS_1 = TextureKey.of("contents_1");
     public static final TextureKey CONTENTS_2 = TextureKey.of("contents_2");
     public static final TextureKey CONTENTS_3 = TextureKey.of("contents_3");