+++ src/client/java/io/github/drakonkinst/worldsinger/datagen/ModTextureMaps.java
@@ -36,12 +36,13 @@ public final class ModTextureMaps {
 
     public static TextureMap aluminumCauldron(Identifier content) {
         return new TextureMap().put(TextureKey.PARTICLE,
-                        TextureMap.getSubId(Blocks.CAULDRON, "_side"))
+                        TextureMap.getSubId(ModBlocks.ALUMINUM_CAULDRON, "_side"))
                 .put(TextureKey.SIDE, TextureMap.getSubId(ModBlocks.ALUMINUM_CAULDRON, "_side"))
-                .put(TextureKey.TOP, TextureMap.getSubId(Blocks.CAULDRON, "_top"))
+                .put(TextureKey.TOP, TextureMap.getSubId(ModBlocks.ALUMINUM_CAULDRON, "_top"))
                 .put(TextureKey.BOTTOM, TextureMap.getSubId(Blocks.CAULDRON, "_bottom"))
-                .put(TextureKey.INSIDE, TextureMap.getSubId(Blocks.CAULDRON, "_inner"))
-                .put(ModTextureKeys.ALUMINUM, TextureMap.getId(ModBlocks.ALUMINUM_BLOCK))
+                .put(TextureKey.INSIDE, TextureMap.getSubId(ModBlocks.ALUMINUM_CAULDRON, "_inner"))
+                .put(ModTextureKeys.SIDE_INNER,
+                        TextureMap.getSubId(ModBlocks.ALUMINUM_CAULDRON, "_side_inner"))
                 .put(TextureKey.CONTENT, content);
     }